backend/ml/document_extractor/requirements.txt
Pillow==10.4.0
pytesseract==0.3.10
PyPDF2==3.0.1
pdf2image==1.17.0